(3p) LSC: Remove `Lint as` gpylint comment directives.
This removes the `# [Lint as:] python{2,3}` header comments that used to tell
gpylint how to parse the file. Gpylint no longer supports that, so these
comments are noops that can be removed.

More information: go/remove-lint-as-lsc and b/181807094

BEGIN_PUBLIC
Remove unused comments related to Python 2 compatibility.
END_PUBLIC

Tested:
    TAP found no affected targets. No targets were built or tested.
    http://test/OCL:441673008:BASE:441674394:1649920083292:5e6420eb
PiperOrigin-RevId: 441831292